+++ backend/src/utils/loader.go
@@ -4,7 +4,6 @@ import (
 	"database/sql"
 	"log"
 	"os"
-
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/joho/godotenv"
 )
